refactor(kafka): migrate Kafka integration from folio-integration-kafka to folio-kafka-producer
