+++ pom.xml
@@ -26,7 +26,7 @@
     <openapi-tools.jackson-databind-nullable.version>0.2.10</openapi-tools.jackson-databind-nullable.version>
     <folio-spring-support.version>10.0.0</folio-spring-support.version>
     <folio-java-checkstyle.version>1.2.0</folio-java-checkstyle.version>
-    <applications-poc-tools.version>4.0.0-SNAPSHOT</applications-poc-tools.version>
+    <applications-poc-tools.version>4.1.0-SNAPSHOT</applications-poc-tools.version>
     <commons-lang3.version>3.20.0</commons-lang3.version>
     <mod-login-keycloak.yaml-file>${project.basedir}/src/main/resources/swagger.api/mod-login-keycloak.yaml
     </mod-login-keycloak.yaml-file>
@@ -86,7 +86,7 @@
 
     <dependency>
       <groupId>org.folio</groupId>
-      <artifactId>folio-integration-kafka</artifactId>
+      <artifactId>folio-kafka-producer</artifactId>
       <version>${applications-poc-tools.version}</version>
     </dependency>
