+++ src/main/java/org/folio/login/LoginKeycloakApplication.java
@@ -1,13 +1,13 @@
 package org.folio.login;
 
 import org.folio.common.configuration.properties.FolioEnvironment;
-import org.folio.integration.kafka.EnableKafka;
+import org.folio.integration.kafka.producer.EnableKafkaProducer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Import;
 
-@EnableKafka
+@EnableKafkaProducer
 @EnableCaching
 @SpringBootApplication
 @Import(FolioEnvironment.class)
