+++ src/main/java/org/folio/login/integration/kafka/configuration/property/KafkaProperties.java
@@ -2,7 +2,7 @@ package org.folio.login.integration.kafka.configuration.property;
 
 import java.util.List;
 import lombok.Data;
-import org.folio.integration.kafka.FolioKafkaProperties.KafkaTopic;
+import org.folio.integration.kafka.producer.KafkaProducerProperties.KafkaTopic;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
 import org.springframework.stereotype.Component;
